refactor: move common mirror tests function into a dedicated package
